# openapi.yaml
openapi: "3.0.3"
info:
  title: netdata-postgres-mcp
  description: >
    Sidecar service that collects Netdata metrics into PostgreSQL and exposes
    them via MCP (Model Context Protocol) over HTTP/SSE.
  version: "1.3.0"
  license:
    name: GPL-3.0-or-later
    url: https://www.gnu.org/licenses/gpl-3.0.html

servers:
  - url: http://localhost:8765
    description: Default local instance

paths:
  /healthz:
    get:
      summary: Liveness probe
      description: Returns 200 if the process is alive.
      operationId: healthz
      tags: [Health]
      responses:
        "200":
          description: Service is alive
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  version:
                    type: string
                    example: 1.3.0

  /readyz:
    get:
      summary: Readiness probe
      description: Returns 200 if the database is reachable, 503 otherwise.
      operationId: readyz
      tags: [Health]
      responses:
        "200":
          description: Service is ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ready
        "503":
          description: Database unreachable
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  reason:
                    type: string
                    example: database unreachable

  /metrics:
    get:
      summary: Prometheus metrics
      description: Returns internal runtime metrics in Prometheus exposition format.
      operationId: metrics
      tags: [Observability]
      responses:
        "200":
          description: Prometheus text format
          content:
            text/plain:
              schema:
                type: string

  /sse:
    get:
      summary: MCP SSE transport
      description: >
        Server-Sent Events endpoint for MCP communication. Clients connect here
        to receive streaming MCP responses. Requires Authorization header if
        MCP_AUTH_TOKEN is configured.
      operationId: mcpSSE
      tags: [MCP]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: SSE stream established
          content:
            text/event-stream:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"

  /message:
    post:
      summary: MCP message endpoint
      description: >
        Submit MCP tool call requests. Used in conjunction with the /sse endpoint
        for bidirectional MCP communication.
      operationId: mcpMessage
      tags: [MCP]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: MCP JSON-RPC request
      responses:
        "200":
          description: MCP response
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/v1/write:
    post:
      summary: Prometheus remote-write receiver
      description: >
        Accepts metrics via the Prometheus remote-write protocol (snappy-compressed
        protobuf). Allows external Prometheus instances to push metrics into the store.
      operationId: remoteWrite
      tags: [Ingest]
      requestBody:
        required: true
        content:
          application/x-protobuf:
            schema:
              type: string
              format: binary
      responses:
        "204":
          description: Samples ingested successfully
        "400":
          description: Invalid payload
        "500":
          description: Ingest error

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Bearer token authentication. Set via MCP_AUTH_TOKEN environment variable.
        Can also be passed as ?token= query parameter for SSE clients.

  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "unauthorized — provide Authorization: Bearer <token>"

tags:
  - name: Health
    description: Kubernetes-compatible health probes
  - name: Observability
    description: Internal metrics and monitoring
  - name: MCP
    description: Model Context Protocol endpoints
  - name: Ingest
    description: Metric ingestion endpoints